Fix: Updates latest migration to unblock future ones.
The migration was meant originally to delete some unused tables from the
permit schema.

When run on non local environments this migration fails because the
sequence in the `permit.x_users` table is shared with the `idm.users`
table.

So this fix creates a new sequence for `idm.users` which then allows the
other tables to be deleted.

For safety the new sequence starts at 100000.