Apply consistent CI workflow
https://github.com/DEFRA/water-abstraction-team/issues/4

As a team we are currently maintaining 11 repos and that's not including documentation, support and deployment related stuff! What we inherited was not consistent; different CI tools used for the same purpose and different CI steps.

An effort was made to get this under control with [water-abstraction-orchestration](https://github.com/DEFRA/water-abstraction-orchestration). But to get a handle on it in such a way as the current team can manage, we're simplifying still further.

This change updates `.github/workflows/ci.yml` to use a template we're applying across all the repos, amended to the requirements of the project. In the future we hope to return to the lessons **water-abstraction-orchestration** taught us, if only to remove the duplication across the projects. But for now the template hopefully simplifies and makes consistent our build process.