Add SonarCloud properties file
Previously, the CI specified the values SonarCloud needs in its steps. But there are a bunch if other things we can tell SonarCloud about which makes managing SonarCloud much easier.

So, we use `sonar-project.properties` files to set what the SonarCloud analysis step needs and what's useful in its UI.