Fix code coverage reporting
Clearly a copy & paste issue broke code coverage reporting. We also switch to a repo agnostic version of the sed command thanks to finding a recent response in the original post that provided a solution.

Finally we tweak the properties file to correctly include the source code files including those in the root of the project when it comes to resolving lcov paths against the project.