Remove snyk dev dependency
https://github.com/DEFRA/water-abstraction-team/issues/6

We use [Snyk](https://snyk.io/) in our CI process to prevent us adding vulnerable dependencies and let us know if anything we do have is an issue.

It looks like the previous team also added the [Snyk package](https://www.npmjs.com/package/snyk) as a dev dependency. However, you only really need that in situations where you can't use the web based version.

As part of cleaning up the repo we're getting rid of this dependency.

We're also getting rid of it's ignore file. There is no reason we should be carrying vulnerable dependencies so we want to know about them.