Add shai-hulud-detect to CI as daily workflow
With the [announcement that the npm supply-chain Shai-Hulud attack is back](https://www.koi.ai/incident/live-updates-sha1-hulud-the-second-coming-hundred-npm-packages-compromised) we're being asked yet again to confirm if our services have been compromised.

The problem is, as a self-replicating worm which npm packages are compromised is a moving target.

Some of our fantastic colleagues have pulled together bash scripts that can be run to test whether a project may have been compromised. Initially, they depended on self-maintained CSV files of compromised packages, but later iterations were pulling down [compromised-packages.txt](https://github.com/Cobenian/shai-hulud-detect/blob/main/compromised-packages.txt).

We'd already found [shai-hulud-detect](https://github.com/Cobenian/compromised-packages.txt) when looking for something 'off-the-shelf', so figured if we're using their list, why not also use their scanner?

We initially looked to include the scan as a CI step. But clearly **shai-hulud-detect** is doing something more than scanning `package-lock.json` files for matches as it took 13 minutes to complete!

So, instead we're going to try adding it as a new GitHub workflow that we schedule to run daily. That way we can automate a more thorough check of our projects.

> NOTE - This is the first project we are adding the workflow too, and where we'll try to iron out any issues.